third_party/WebKit/LayoutTests/http/tests/inspector/command-line-api-inspect.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="console-test.js"></script>
<script>

function onload()
{
    runTest();
}

function test()
{
    InspectorTest.addSniffer(WebInspector, "inspect", sniffInspect, true);

    function sniffInspect(objectId, hints)
    {
        InspectorTest.addResult("WebInspector.inspect called with: " + objectId.description);
        InspectorTest.addResult("WebInspector.inspect's hints are: " + JSON.stringify(Object.keys(hints)));
    }

    function evalAndDump(expression, next)
    {
        InspectorTest.addResult("\n");
        InspectorTest.evaluateInConsole(expression, dumpCallback);
        function dumpCallback(text)
        {
            text = text.replace(/Test of session.*/, "Test of session storage...");
            InspectorTest.addResult(expression + " = " + text);
            if (next)
                next();
        }
    }

    InspectorTest.runTestSuite([
        function testRevealElement(next)
        {
            InspectorTest.addSniffer(WebInspector.Revealer, "reveal", step2);
            evalAndDump("inspect($('#p1'))");

            function step2(node)
            {
                if (!(node instanceof WebInspector.DOMNode))
                    return;

                InspectorTest.addResult("Selected node id: '" + WebInspector.panels.elements.selectedDOMNode().getAttribute("id") + "'.");
                next();
            }
        }
    ]);
}

</script>
</head>

<body onload="onload()">
<p id="p1">
Tests that inspect() command line api works.
</p>

</body>
</html>
